+++ docx4j-core/src/main/java/org/docx4j/openpackaging/parts/SpreadsheetML/CustomDataPart.java
@@ -1,0 +1,51 @@
+/*
+ *  Copyright 2026, Plutext Pty Ltd.
+ *
+ *  This file is part of docx4j.
+
+    docx4j is licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+        http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+
+ */
+package org.docx4j.openpackaging.parts.SpreadsheetML;
+
+import org.docx4j.openpackaging.contenttype.ContentType;
+import org.docx4j.openpackaging.contenttype.ContentTypes;
+import org.docx4j.openpackaging.exceptions.InvalidFormatException;
+import org.docx4j.openpackaging.parts.PartName;
+import org.docx4j.openpackaging.parts.WordprocessingML.BinaryPart;
+import org.docx4j.openpackaging.parts.relationships.Namespaces;
+
+/**
+ * An add-in's custom data ([MS-XLSX] 2.1.2), opaque bytes (content type application/binary),
+ * related from its CustomDataPropertiesPart.
+ *
+ * @since 17.1.1 (CR-022)
+ */
+public class CustomDataPart extends BinaryPart {
+
+	public CustomDataPart(PartName partName) throws InvalidFormatException {
+		super(partName);
+		init();
+	}
+
+	public CustomDataPart() throws InvalidFormatException {
+		super(new PartName("/xl/customData/customData1.dat"));
+		init();
+	}
+
+	public void init() {
+		setContentType(new ContentType(ContentTypes.SPREADSHEETML_CUSTOM_DATA));
+		setRelationshipType(Namespaces.SPREADSHEETML_CUSTOM_DATA);
+	}
+
+}